requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
coverage==6.1.1
cssselect==1.1.0
Django==3.2.9
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-shortcuts==1.6
flake8==4.0.1
gunicorn==20.1.0
iniconfig==1.1.1
lxml==4.6.3
mccabe==0.6.1
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
pyquery==1.4.3
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
